Split off config.go
